[Http] Move Provider class to abstract namespace.
